+++ packages/palette/.storybook/preview.js
@@ -1,40 +1,57 @@
-import React from "react"
-import { addDecorator } from "@storybook/react"
+import React, { useState } from "react"
 import { Theme } from "../src/Theme"
+import { THEMES } from "../src/themes"
 import { injectGlobalStyles } from "../src/helpers/injectGlobalStyles"
 import { INITIAL_VIEWPORTS } from "@storybook/addon-viewport"
 import { breakpoints } from "../src/Theme"
+import { Tabs, Tab } from "../src/elements/Tabs"
 
 const { GlobalStyles } = injectGlobalStyles()
 
-addDecorator((storyFn) => (
-  <Theme>
-    <>
-      <GlobalStyles />
-      {storyFn()}
-    </>
-  </Theme>
-))
+export const decorators = [
+  (Story) => {
+    const [theme, setTheme] = useState("v2")
 
-const viewports = Object.entries(breakpoints).reduce(
-  (memo, [key, width]) => {
-    if (['__docgenInfo', 'displayName'].includes(key)) {
-      return memo
-    }
+    return (
+      <Theme theme={THEMES[theme]}>
+        <>
+          <GlobalStyles />
 
-    return {
-      ...memo,
-      [key]: {
-        name: `breakpoint: ${key}`,
-        styles: {
-          width,
-          height: "100%",
-        },
-      },
-    }
+          <Tabs
+            onChange={({ name }) => {
+              setTheme(name)
+            }}
+          >
+            <Tab name="v2">
+              <Story />
+            </Tab>
+
+            <Tab name="v3">
+              <Story />
+            </Tab>
+          </Tabs>
+        </>
+      </Theme>
+    )
   },
-  {}
-)
+]
+
+const viewports = Object.entries(breakpoints).reduce((memo, [key, width]) => {
+  if (["__docgenInfo", "displayName"].includes(key)) {
+    return memo
+  }
+
+  return {
+    ...memo,
+    [key]: {
+      name: `breakpoint: ${key}`,
+      styles: {
+        width,
+        height: "100%",
+      },
+    },
+  }
+}, {})
 
 export const parameters = {
   options: {
